Add new folder to separate game and auth and change html
Created two separate folders in order to compartmentalize the different
aspects of the game and the auth for login. Also added an events, api,
and ui file to each of the new folders. Then Made basic changes to the
HTML file to represent a tic-tac-toe grid. Also added an event listener
to the newly created grid and am in the process of figuring out how to
log the player moves to the grid using the event listener.