Correct player change function bug
Made a change to the onPlayerMove function on the events.js file.
Previously if you selected an invalid spot on the board it would still
switch to the next player. Now the player will only change when a valid
move is registered and if an invalid spot is detected the player
remains the same.