Add function to log player move to array
Added a function to the events.js file and a separate event handler to
the app.js file in order to log player moves to an array. All player
moves are now logged to an array representing the boxes on the
tic-tac-toe board and once a spot on the array has been filled it cannot
be overwritten. The next step will now be to check the logged moves for
a winning combo.