Add function to determine end game and button for new game
Made corrections to previous functions of currentGameBoard and
onPlayerMove so that they now pass through an isGameOver function first.
So that once a game is over no more moves can be logged to the board or
the currentBoard array.

Added a button to the html in order to trigger a new game. Added an
event listener to app.js for the new game button. Also added a function
to events.js that resets all game info when a new game is triggered.